backend/requirements-frozen.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioredis==2.0.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.14.0
alphashape==1.3.1
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncio-throttle==1.0.2
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
black==23.11.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-log==0.4.0
click-plugins==1.1.1.2
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
Crawl4AI==0.6.0
cryptography==45.0.5
cssselect==1.3.0
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.116.1
feedparser==6.0.10
filelock==3.18.0
flake8==6.1.0
frozenlist==1.7.0
fsspec==2025.7.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.44
gnews==0.4.0
gotrue==2.12.3
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
isort==5.12.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kaitaistruct==0.10
kombu==5.4.2
lark==1.2.2
Levenshtein==0.27.1
limits==5.4.0
litellm==1.53.3
lxml==5.3.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
mypy==1.7.1
mypy_extensions==1.1.0
narwhals==1.47.1
networkx==3.5
newsapi-python==0.2.7
nltk==3.9.1
numpy==1.26.4
openai==1.59.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
passlib==1.7.4
pathspec==0.12.1
pillow==10.4.0
pinecone==5.0.0
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
playwright==1.53.0
plotly==5.17.0
pluggy==1.6.0
postgrest==1.1.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
pyarrow==21.0.0
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.6.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
PySocks==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-json-logger==2.0.7
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
realtime==2.6.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
rtree==1.4.0
safetensors==0.5.3
schedule==1.2.0
scikit-learn==1.6.0
scipy==1.16.0
selenium==4.16.0
selenium-wire==5.1.0
sentence-transformers==3.3.1
setuptools==80.9.0
sgmllib3k==1.0.0
shapely==2.1.1
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.36
starlette==0.47.1
storage3==0.12.0
streamlit==1.28.2
StrEnum==0.4.15
structlog==23.2.0
supabase==2.17.0
supafunc==0.10.1
sympy==1.14.0
tenacity==8.5.0
tf-playwright-stealth==1.2.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
trimesh==4.7.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
validators==0.35.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
